Crate azalea_client

Significantly abstract azalea_protocol so it’s actually useable for real clients. If you want to make bots, you should use the azalea crate instead.

Re-exports

pub use client::start_ecs_runner;

pub use events::Event;

pub use movement::PhysicsState;

pub use movement::SprintDirection;

pub use movement::StartSprintEvent;

pub use movement::StartWalkEvent;

pub use movement::WalkDirection;

Modules

attack

chat

Implementations of chat-related features.

chunks

Used for Minecraft’s chunk batching introduced in 23w31a (1.20.2). It’s used for making the server spread out how often it sends us chunk packets depending on our receiving speed.

configuration

disconnect

Disconnect a client from the server.

events

Defines the Event enum and makes those events trigger when they’re sent in the ECS.

interact

inventory

mining

movement

packet_handling

ping

Ping Minecraft servers.

raw_connection

respawn

task_pool

Borrowed from bevy_core.

Structs

Account

Something that can join Minecraft servers.

Client

Client has the things that a user interacting with the library will want.

ClientInformation

A component that contains some of the “settings” for this client that are sent to the server, such as render distance. This is only present on local players.

DefaultPlugins

This plugin group will add all the default plugins necessary for Azalea to work.

GameProfileComponent

A component only present in players that contains the GameProfile (which you can use to get a player’s name).

Hunger

InstanceHolder

A component that keeps strong references to our PartialInstance and Instance for local players.

JoinedClientBundle

A bundle for the components that are present on a local player that is currently in the game protocol state. If you want to filter for this, just use LocalEntity.

PlayerInfo

A player in the tab list.

StartClientOpts

TabList

A component that contains a map of player UUIDs to their information in the tab list.

TickBroadcast

A resource that contains a [broadcast::Sender] that will be sent every Minecraft tick.

Enums

AccountOpts

The parameters that were passed for creating the associated Account.

JoinError

An error that happened while joining the server.