azalea_client
Module movement

Structs

KnockbackEvent

An event sent by the server that sets or adds to our velocity. Usually KnockbackKind::Set is used for normal knockback and KnockbackKind::Add is used for explosions, but some servers (notably Hypixel) use explosions for knockback.

LastSentLookDirection

A component that contains the look direction that was last sent over the network.

MoveEventsSet

PhysicsState

Component for entities that can move and sprint. Usually only in LocalEntitys.

PlayerMovePlugin

StartSprintEvent

An event sent when the client starts sprinting. This does not get sent for non-local entities.

StartWalkEvent

An event sent when the client starts walking. This does not get sent for non-local entities.

Enums

KnockbackType

MovePlayerError

SprintDirection

The directions that we can sprint in. It’s a subset of WalkDirection.

WalkDirection

Functions

handle_knockback

handle_sprint

The system that makes the player start sprinting when they receive a StartSprintEvent.

handle_walk

The system that makes the player start walking when they receive a StartWalkEvent.

local_player_ai_step

Makes the bot do one physics tick. Note that this is already handled automatically by the client.

send_position