azalea_client::packet_handling
Module game

Structs

AddPlayerEvent

A player joined the game (or more specifically, was added to the tab list of a local player).

DeathEvent

Event for when an entity dies. dies. If it’s a local player and there’s a reason in the death screen, the ClientboundPlayerCombatKillPacket will be included.

InstanceLoadedEvent

An instance (aka world, dimension) was loaded by a client.

KeepAliveEvent

A KeepAlive packet is sent from the server to verify that the client is still connected.

PacketEvent

An event that’s sent when we receive a packet.

RemovePlayerEvent

A player left the game (or maybe is still in the game and was just removed from the tab list of a local player).

ResourcePackEvent

SendPacketEvent

An event for sending a packet to the server while we’re in the game state.

UpdatePlayerEvent

A player was updated in the tab list of a local player (gamemode, display name, or latency changed).

Functions

handle_send_packet_event

process_packet_events

send_packet_events